refactor(ansi_colors): extracted actual ANSI color codes generation
Moved ANSI color codes generation to separate interface, so it can be easily extended.
